+++ bundles/analysis/bundle/analyse/locale/en.js
@@ -20,19 +20,23 @@ Oskari.registerLocalization({
                         "id": "oskari_analyse_buffer",
                         "label": "Buffer",
                         "classForMethod": "buffer",
-                        "selected": true
+                        "selected": true,
+                        "tooltip": ""
                     }, {
                         "id": "oskari_analyse_aggregate",
                         "label": "Aggregate",
-                        "classForPreview": "aggregate"
+                        "classForPreview": "aggregate",
+                        "tooltip": ""
                     }, {
                         "id": "oskari_analyse_union",
                         "label": "Union",
-                        "classForPreview": "union"
+                        "classForPreview": "union",
+                        "tooltip": ""
                     }, {
                         "id": "oskari_analyse_intersect",
                         "label": "Intersect",
-                        "classForPreview": "intersect"
+                        "classForPreview": "intersect",
+                        "tooltip": ""
                     }
                 ]
             },
@@ -134,6 +138,7 @@ Oskari.registerLocalization({
                 "values": {
                     "title": "Filter",
                     "placeholders": {
+                        "case-sensitive": "",
                         "attribute": "Attribute",
                         "boolean": "Logical operator",
                         "operator": "Oprator",